move JwtProperties to core project
