--- backend/cms-backend/core/src/main/java/ru/yakovlev05/cms/core/CoreApplication.java
@@ -1,13 +1,0 @@
-//package ru.yakovlev05.cms.core;
-//
-//import org.springframework.boot.SpringApplication;
-//import org.springframework.boot.autoconfigure.SpringBootApplication;
-//
-//@SpringBootApplication
-//public class CoreApplication {
-//
-//	public static void main(String[] args) {
-//		SpringApplication.run(CoreApplication.class, args);
-//	}
-//
-//}
