+++ backend/cms-backend/core/src/main/java/ru/yakovlev05/cms/core/security/JwtProvider.java
@@ -1,0 +1,4 @@
+package ru.yakovlev05.cms.core.security;
+
+public class JwtProvider {
+}
