+++ backend/cms-backend/core/src/main/resources/application.yaml
@@ -1,3 +1,10 @@
 spring:
   application:
     name: core
+
+
+
+jwt:
+  secret: ${JWT_SECRET:only_test_123456789123456789}
+  accessTokenValidityInMs: ${ACCESS_TOKEN_VALIDITY_IN_MS:86400000}
+  refreshTokenValidityInMs: ${REFRESH_TOKEN_VALIDITY_IN_MS:2592000000}
