+++ backend/cms-backend/microservice-auth/src/main/java/ru/yakovlev05/cms/auth/MicroserviceAuthApplication.java
@@ -3,7 +3,7 @@ package ru.yakovlev05.cms.auth;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "ru.yakovlev05.cms")
 public class MicroserviceAuthApplication {
 
     public static void main(String[] args) {
