+++ backend/cms-backend/microservice-auth/src/main/resources/application.yaml
@@ -18,8 +18,3 @@ spring:
     producer:
       key-serializer: org.apache.kafka.common.serialization.StringSerializer
       value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
-jwt:
-  secret: ${JWT_SECRET:only_test_123456789123456789}
-  accessTokenValidityInMs: ${ACCESS_TOKEN_VALIDITY_IN_MS:86400000}
-  refreshTokenValidityInMs: ${REFRESH_TOKEN_VALIDITY_IN_MS:2592000000}
-
