+++ backend/cms-backend/microservice-catalog/src/main/java/ru/yakovlev05/cms/catalog/MicroserviceCatalogApplication.java
@@ -3,7 +3,7 @@ package ru.yakovlev05.cms.catalog;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
-@SpringBootApplication
+@SpringBootApplication(scanBasePackages = "ru.yakovlev05.cms")
 public class MicroserviceCatalogApplication {
 
     public static void main(String[] args) {
