--- backend/cms-backend/microservice-catalog/src/main/java/ru/yakovlev05/cms/catalog/props/JwtProperties.java
@@ -1,12 +1,0 @@
-package ru.yakovlev05.cms.catalog.props;
-
-import lombok.Data;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@Data
-@ConfigurationProperties(prefix = "jwt")
-public class JwtProperties {
-    private String secret;
-}
