+++ backend/cms-backend/microservice-catalog/src/main/resources/application.yaml
@@ -17,9 +17,6 @@ spring:
     enabled: true
     change-log: classpath:/db/changelog/db.changelog-master.xml
 
-jwt:
-  secret: ${JWT_SECRET:only_test_123456789123456789}
-
 s3:
   accessKey: ${S3_ACCESS_KEY:user}
   secretKey: ${S3_SECRET_KEY:password}
